requirements.txt
Adafruit-Blinka==8.31.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-dht==4.0.4
adafruit-circuitpython-requests==2.0.4
adafruit-circuitpython-typing==1.9.6
Adafruit-PlatformDetect==3.59.0
Adafruit-PureIO==1.1.11
awscrt==0.20.2
awsiotsdk==1.0.0.dev0
bidict==0.22.1
blinker==1.7.0
boto3==1.34.31
botocore==1.34.31
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.5.0
eventlet==0.35.1
Flask==3.0.1
Flask-SocketIO==5.3.6
gevent==23.9.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.4
paho-mqtt==1.6.1
pyftdi==0.55.0
pyserial==3.5
python-dateutil==2.8.2
python-engineio==4.8.2
python-socketio==5.11.0
pytz==2023.4
pyusb==1.2.1
requests==2.31.0
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
s3transfer==0.10.0
simple-websocket==1.0.0
six==1.16.0
sysv-ipc==1.1.0
typing_extensions==4.9.0
urllib3==2.0.7
websocket-client==1.7.0
Werkzeug==3.0.1
wsproto==1.2.0
zope.event==5.0
zope.interface==6.1
